Derive HeapHandle from SafeBuffer. Change buffer methods to take ulong to allow working with large buffers. Remove the ability to shrink buffer sizes to reduce reallocations.

+++ src/xTask/Interop/StreamBuffer.cs
@@ -88,8 +88,8 @@ namespace XTask.Interop
         {
             Debug.Assert(size >= 0);
 
-            if (this.stream != null && this.buffer.Capacity >= size) return;
-            this.buffer.EnsureCapacity(size);
+            if (this.stream != null && this.buffer.Capacity >= (ulong)size) return;
+            this.buffer.EnsureCapacity((ulong)size);
 
             long oldLength = this.Length;
             long oldPosition = this.Position;
